BaseModel to inherit from
